requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sentence-transformers==3.4.1
chromadb==0.6.3
python-pptx==1.0.2
pdfplumber==0.11.5
python-dotenv==1.0.1
requests==2.32.3
pydantic-settings==2.7.1
python-multipart==0.0.20
openai==1.61.0
pdf2image==1.17.0
pytesseract==0.3.13
alembic==1.17.0
psycopg2-binary==2.9.11
sqlmodel==0.0.27
boto3==1.36.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
arq==0.25.0
redis==5.0.0
python-json-logger==2.0.7
dspy-ai==2.5.43
